[New Physical IO] Migrating new design with Blob (#294)
This PR merges the new PhysicalIO changes to the Blob object and start
to use the new implementation.

Next Steps:
- Range coalescing implementation
- Retry policy implementation

PR History:

https://github.com/awslabs/analytics-accelerator-s3/pull/286
https://github.com/awslabs/analytics-accelerator-s3/pull/287
https://github.com/awslabs/analytics-accelerator-s3/pull/288
https://github.com/awslabs/analytics-accelerator-s3/pull/289

existing APIs or behaviors?
No

No
Unit test

n/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).